refactor(core): isolate objective evidence state
Why:
- Transient routing and pre-binding extraction values were embedded in the objective service, obscuring their application-layer ownership.
- Unused `paper_id` and `anchor_ids` aliases prolonged silent compatibility paths with no active producers.

What:
- Move EvidenceCandidate and ExtractedEvidenceDraft into dedicated objective routing and extraction modules and update direct imports and ownership docs.
- Remove legacy aliases from paper frames, evidence candidates, and extraction drafts, with regression coverage for canonical fields.
- Remove obsolete transient conversion helpers and domain imports from the objective service.

Impact:
- Objective HTTP, persistence, pipeline, and durable domain behavior remain unchanged; internal payloads now require `document_id` and `evidence_anchor_ids`.
- No compatibility wrapper or new runtime abstraction is introduced.

Tests:
- `./.venv/bin/python -m pytest tests/unit/services/test_research_objective_service.py tests/unit/application/test_objective_transient_models.py tests/unit/application/test_objective_analysis_service.py tests/unit/application/test_collection_build_pipeline_runner.py tests/integration/services/test_task_runner.py -q` (123 passed)
- `./.venv/bin/python -m compileall -q application tests`
- `python3 scripts/check_docs_governance.py`
- `git diff --cached --check`

Refs:
- n/a
